Rename dg_effect.c
