+++ include/dg_effect.h
