+++ main.lsf
@@ -189,7 +189,7 @@ Overlay OVY_29
 	Object asm/overlay_29_022E1A40.o
 	Object src/dg_camera_022E32E8.o
 	Object asm/overlay_29_022E330C.o
-	Object src/overlay_29_022E37B8.o
+	Object src/dg_effect_022E37B8.o
 	Object asm/overlay_29_022E37DC.o
 	Object src/overlay_29_022E95F4.o
 	Object asm/overlay_29_022E9618.o
