+++ src/dg_effect_022E37B8.c
@@ -1,4 +1,4 @@
-#include "overlay_29_022E37B8.h"
+#include "dg_effect.h"
 
 u8 EntityIsValid__022E37B8(struct entity *entity)
 {
